Working Prototype Without ReservationOptions Class
A lot was added here, it is hastily put together but the system works without ReservationOptions for now, and must be changed once that class is pushed.

Added:
images for each room type

created rooms file info

rooms page now lists correct info with
correct image attached

fixed manipfile to read room file and populate rooms list

made cust list into a linked list

fixed manipfile to read cust file and populate cust list

customer class currently has check in/out variables and a room num variable just to get things working, this must be changed later

changed some formatting for the reports so that it shows the correct dates

added room number to the room info panel

changed review reservation so that it only shows a review if the information matches a customer on file. error message is shown if input does not validate and user is taken back to options menu